Ajout touche action + coffres ouvrables - Refactoring

+++ src/main/java/com/mesi/panels/maps/MapGenerator.java
@@ -22,7 +22,7 @@ public class MapGenerator {
     /**********  Methods  **********/
 
     public void map_1() throws IOException {
-        MapModel map = new Map_1(90, 50, Game.characterCoordinates[0] / Constant.TILE_SIZE, Game.characterCoordinates[1] / Constant.TILE_SIZE, (Integer) Game.direction.get(0));
+        MapModel map = new Map_1(90, 50, Game.characterCoordinates[0] / Constant.TILE_SIZE, Game.characterCoordinates[1] / Constant.TILE_SIZE, Game.direction.get(0));
         String name = "MAP_1";
 
         MainZeldo.mapList.put(name, map);
@@ -30,7 +30,7 @@ public class MapGenerator {
 
     public void map_2() throws IOException {
         //MapModel map = new Map_2(40, 24, Game.teleportPositionX, Game.teleportPositionY, (Integer) Game.direction.get(0));
-        MapModel map = new Map_2(Constant.FRAME_WIDTH / 32, Constant.FRAME_HEIGHT / 32, Game.characterCoordinates[0] / Constant.TILE_SIZE, Game.characterCoordinates[1] / Constant.TILE_SIZE, (Integer) Game.direction.get(0));
+        MapModel map = new Map_2(Constant.FRAME_WIDTH / 32, Constant.FRAME_HEIGHT / 32, Game.characterCoordinates[0] / Constant.TILE_SIZE, Game.characterCoordinates[1] / Constant.TILE_SIZE, Game.direction.get(0));
         String name = "MAP_2";
 
         MainZeldo.mapList.put(name, map);
